直播课堂逻辑修改

+++ code/db/src/main/java/com/db/excellentCourses/HourClassDao.java
@@ -40,7 +40,7 @@ public class HourClassDao extends BaseDao {
     }
 
     //修改价格
-    public void updatePriceEntry(ObjectId id,int price){
+    public void updatePriceEntry(ObjectId id,double price){
         BasicDBObject query = new BasicDBObject(Constant.ID,id);
         BasicDBObject updateValue=new BasicDBObject(Constant.MONGO_SET,new BasicDBObject("cnp",price));
         update(MongoFacroty.getAppDB(), Constant.COLLECTION_HOUR_CLASS, query,updateValue);
